+++ whatsapp_addon/custom_component/__init__.py
@@ -17,6 +17,11 @@ async def async_setup(hass: HomeAssistant, config: ConfigType) -> bool:
     async def send_message(call: ServiceCall) -> None:
         await hass.async_add_executor_job(whatsapp.send_message, call.data)
 
+    @callback
+    async def set_status(call: ServiceCall) -> None:
+        await hass.async_add_executor_job(whatsapp.set_status, call.data)
+
     hass.services.async_register(DOMAIN, 'send_message', send_message)
+    hass.services.async_register(DOMAIN, 'set_status', set_status)
 
     return True