+++ whatsapp_addon/custom_component/whatsapp.py
@@ -6,3 +6,6 @@ HOST = 'http://{{HOSTNAME}}:3000/'
 class Whatsapp:
     def send_message(self, data):
         return requests.post(url_normalize(f'{HOST}/sendMessage'), json=data).content == 'OK'
+
+    def set_status(self, data):
+        return requests.post(url_normalize(f'{HOST}/setStatus'), json=data).content == 'OK'