+++ whatsapp_addon/CHANGELOG.md
@@ -1,5 +1,12 @@
 <!-- https://developers.home-assistant.io/docs/add-ons/presentation#keeping-a-changelog -->
 
+## 1.0.2
+
+- Addedd message revoke event.
+- Added buttons message type (view documentation) (may not work properly on some devices)
+- Added set status service (for sets the current user's status message)
+- Bug fixed.
+
 ## 1.0.1
 
 - Initial release